requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
pymongo==4.7.1
python-dotenv==1.0.1
plotly==5.21.0
